# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tradosphere OS Gateway API
  version: "1.0.0"
  description: |-
    Unified API gateway for Tradosphere OS (Sprint 9, task 9.4). Implements
    the architecture decided in full in Decision D19 (EXECUTION_BOOK.md)
    before any gateway code was written.

    Two route families exist side by side:

    - Proxied (Auth, Market Data, Education, Portfolio, Analytics): the
      gateway reverse-proxies to each service's existing HTTP API via a
      hand-rolled wildcard handler that forwards the Authorization header
      verbatim. Each upstream service remains the sole auth authority for
      its own routes -- the gateway does NOT re-authenticate proxied
      requests.
    - In-process (Research, AI Council, CIO, Paper Trading, Journal): these
      five services are library-only with no HTTP surface of their own, so
      their routes are implemented directly on the gateway and authenticated
      by the gateway's own requireAuth (identical factory pattern to every
      other service in this repo).

    All domain routes are versioned under /v1/<service>/.... Infra routes
    (/health, /health/services, /metrics, /stream, /documentation,
    /openapi.yaml) are unversioned and, along with auth's own signup/login,
    are the only unauthenticated routes in the gateway.
servers:
  - url: http://localhost:4000
    description: Local gateway (apps/api)
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/signup:
    post:
      operationId: authSignup
      summary: Create a new account and issue an access/refresh token pair
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '409':
          description: Email already in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded (Redis-backed, global to services/auth).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/portfolio/positions:
    get:
      operationId: portfolioListPositions
      summary: Current open positions, netted per symbol
      description: |-
        Computed fresh from journal_entries on every call, never cached
        (Decision D17) -- closed trades no longer contribute, and a symbol
        whose open trades net to exactly zero is omitted rather than
        reported as a zero-quantity row.
      tags: [Portfolio]
      responses:
        '200':
          description: The caller's open positions, sorted by symbol.
          content:
            application/json:
              schema:
                type: object
                properties:
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Position'
                required: [positions]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        Every proxied service verifies this itself against its own JWT
        secret; the gateway's own requireAuth (used only by the 10
        in-process routes) uses the identical verifyAccessToken /
        InvalidTokenError contract from @tradosphere/auth as every other
        service.
  schemas:
    ErrorResponse:
      type: object
      description: |-
        The generic shape every service's setErrorHandler falls back to for
        unexpected errors, and the shape most caught domain errors use
        (sometimes with extra fields alongside error, e.g. 409 responses
        that also carry missingPriceSymbols). Not a Fastify-default
        {statusCode,error,message} envelope -- just {error, ...}.
      properties:
        error:
          type: string
      required: [error]
    ValidationFailure:
      type: object
      description: The zod-failure shape validateBody() returns on every 400, identical across every service (reused verbatim from services/education's original).
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required: [error, details]
...
    Role:
      type: string
      enum: [admin, trader, viewer]
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
      required: [email, password]
...
    AuthResult:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            role:
              $ref: '#/components/schemas/Role'
          required: [id, email, role]
      required: [accessToken, refreshToken, user]
...
    PositionDirection:
      type: string
      enum: [long, short]
    Position:
      type: object
      description: Net position for one symbol, computed from open trade records only as a quantity-weighted average entry price (Decision D17).
      properties:
        symbol:
          type: string
        direction:
          $ref: '#/components/schemas/PositionDirection'
        quantity:
          type: number
        averageEntryPrice:
          type: number
      required: [symbol, direction, quantity, averageEntryPrice]
...